requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
decouple==0.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.54.0
django-cors-headers==4.3.0
django-extensions==3.2.3
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
Faker==19.10.0
faker_food==0.2.0
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.28.1
packaging==23.1
pathspec==0.11.2
pdfkit==1.0.0
Pillow==10.0.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
wkhtmltopdf==0.2
yarl==1.9.2
